Added gauge badge
