+++ src/badge/gauge-badge-config.ts
@@ -1,0 +1,7 @@
+import { LovelaceBadgeConfig } from "../ha/lovelace";
+
+export interface ModernCircularGaugeBadgeConfig extends LovelaceBadgeConfig {
+  entity: string;
+  min?: number;
+  max?: number;
+}
