+++ src/ha/lovelace.ts
@@ -1,4 +1,9 @@
 
+export interface LovelaceBadgeConfig {
+    type: string;
+    visibility?: any;
+}
+
 export interface LovelaceLayoutOptions {
     grid_columns?: number | "full";
     grid_rows?: number | "auto";
